# api/docs/swagger.yaml
definitions:
...
  auth_service.AddPermissionScopeRequest:
    properties:
      client_platform_id:
        type: string
      method:
        type: string
      path:
        type: string
      permission_id:
        type: string
    type: object
...
  auth_service.AddRolePermissionRequest:
    properties:
      permission_id:
        type: string
      role_id:
        type: string
    type: object
  auth_service.AddRoleRequest:
    properties:
      client_type_id:
        type: string
      name:
        type: string
    type: object
...
  auth_service.CreatePermissionRequest:
    properties:
      client_platform_id:
        type: string
      name:
        type: string
      parent_id:
        type: string
    type: object
...
  auth_service.GetPermissionListResponse:
    properties:
      count:
        type: integer
      permissions:
        items:
          $ref: '#/definitions/auth_service.Permission'
        type: array
    type: object
...
  auth_service.Permission:
    properties:
      client_platform_id:
        type: string
      id:
        type: string
      name:
        type: string
      parent_id:
        type: string
    type: object
  auth_service.PermissionScope:
    properties:
      client_platform_id:
        type: string
      method:
        type: string
      path:
        type: string
      permission_id:
        type: string
    type: object
  auth_service.PermissionScopePrimaryKey:
    properties:
      client_platform_id:
        type: string
      method:
        type: string
      path:
        type: string
      permission_id:
        type: string
    type: object
...
  auth_service.Role:
    properties:
      client_type_id:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  auth_service.RolePermission:
    properties:
      permission_id:
        type: string
      role_id:
        type: string
    type: object
  auth_service.RolePermissionPrimaryKey:
    properties:
      permission_id:
        type: string
      role_id:
        type: string
    type: object
...
  auth_service.UpdatePermissionRequest:
    properties:
      client_platform_id:
        type: string
      id:
        type: string
      name:
        type: string
      parent_id:
        type: string
    type: object
...
  auth_service.UpdateRoleRequest:
    properties:
      client_type_id:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  auth_service.UpsertScopeRequest:
    properties:
      client_platform_id:
        type: string
      method:
        type: string
      path:
        type: string
    type: object
...
  http.Response:
    properties:
      data:
        type: object
      description:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a api gateway
  license: {}
  termsOfService: https://udevs.io
paths:
...
  /permission:
    get:
      consumes:
      - application/json
      description: Get Permission List
      operationId: get_permission_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetPermissionListResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.GetPermissionListResponse'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Permission List
      tags:
      - Permission
    post:
      consumes:
      - application/json
      description: Create Permission
      operationId: create_permission
      parameters:
      - description: CreatePermissionRequestBody
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/auth_service.CreatePermissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Permission data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.Permission'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Permission
      tags:
      - Permission
    put:
      consumes:
      - application/json
      description: Update Permission
      operationId: update_permission
      parameters:
      - description: UpdatePermissionRequestBody
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/auth_service.UpdatePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Permission data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.Permission'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Permission
      tags:
      - Permission
  /permission-scope:
    delete:
      consumes:
      - application/json
      description: Get PermissionScope
      operationId: delete_permission_scope
      parameters:
      - description: PermissionScopePrimaryKeyBody
        in: body
        name: permission-scope
        required: true
        schema:
          $ref: '#/definitions/auth_service.PermissionScopePrimaryKey'
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete PermissionScope
      tags:
      - PermissionScope
    post:
      consumes:
      - application/json
      description: Create PermissionScope
      operationId: add_permission_scope
      parameters:
      - description: AddPermissionScopeRequestBody
        in: body
        name: permission-scope
        required: true
        schema:
          $ref: '#/definitions/auth_service.AddPermissionScopeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: PermissionScope data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.PermissionScope'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create PermissionScope
      tags:
      - PermissionScope
  /permission/{permission-id}:
    delete:
      consumes:
      - application/json
      description: Get Permission
      operationId: delete_permission
      parameters:
      - description: permission-id
        in: path
        name: permission-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Permission
      tags:
      - Permission
    get:
      consumes:
      - application/json
      description: Get Permission By ID
      operationId: get_permission_by_id
      parameters:
      - description: permission-id
        in: path
        name: permission-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: PermissionBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.Permission'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Permission By ID
      tags:
      - Permission
...
  /role:
    post:
      consumes:
      - application/json
      description: Create Role
      operationId: create_role
      parameters:
      - description: AddRoleRequestBody
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/auth_service.AddRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Role data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.Role'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Role
      tags:
      - Role
    put:
      consumes:
      - application/json
      description: Update Role
      operationId: update_role
      parameters:
      - description: UpdateRoleRequestBody
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/auth_service.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Role data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.Role'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Role
      tags:
      - Role
  /role-permission:
    delete:
      consumes:
      - application/json
      description: Get RolePermission
      operationId: delete_role_permission
      parameters:
      - description: RolePermissionPrimaryKeyBody
        in: body
        name: role-permission
        required: true
        schema:
          $ref: '#/definitions/auth_service.RolePermissionPrimaryKey'
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete RolePermission
      tags:
      - RolePermission
    post:
      consumes:
      - application/json
      description: Create RolePermission
      operationId: add_role_permission
      parameters:
      - description: AddRolePermissionRequestBody
        in: body
        name: role-permission
        required: true
        schema:
          $ref: '#/definitions/auth_service.AddRolePermissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: RolePermission data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.RolePermission'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create RolePermission
      tags:
      - RolePermission
  /role/{role-id}:
    delete:
      consumes:
      - application/json
      description: Get Role
      operationId: delete_role
      parameters:
      - description: role-id
        in: path
        name: role-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Role
      tags:
      - Role
...
  /upsert-scope:
    post:
      consumes:
      - application/json
      description: Upsert Scope
      operationId: upsert_scope
      parameters:
      - description: UpsertScopeRequestBody
        in: body
        name: upsert-scope
        required: true
        schema:
          $ref: '#/definitions/auth_service.UpsertScopeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Role data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.Role'
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Upsert Scope
      tags:
      - UpsertScope
...
swagger: "2.0"